feat: added get_version function to pyscan/general, implemented it in runfinfo, and also created a basic test file to ensure it works as expected.
