+++ test/general/test_get_version.py
@@ -1,0 +1,8 @@
+from pyscan.general.get_version import get_version
+
+
+def test_get_version():
+    version = get_version()
+    assert type(version) is str, "get_version did not return a string"
+    assert version != "no valid version found", "get_version did not find the version file"
+    assert version[0] == 'v', "version string did not start with v"
